requirements.txt
Flask==3.0.3
tensorflow-intel==2.13.0
tensorflow-estimator==2.13.0
- keras==3.10.0
 keras==2.13.1
numpy==1.24.3
Pillow==11.3.0
matplotlib==3.7.5
Werkzeug==3.0.6
gunicorn==21.2.0
